+++ tools/make_docs.sh
@@ -12,7 +12,7 @@ require_java17 javadoc
 
 ROOT_DIR="$(cd "${SCRIPT_DIR}/.." && pwd)"
 OUT_PATH="${ROOT_DIR}/out/docs"
-CLASS_PATH="${ROOT_DIR}/third_party/jogamp/jar/gluegen-rt.jar:${ROOT_DIR}/third_party/jogamp/jar/jogl-all.jar"
+CLASS_PATH="${ROOT_DIR}/third_party/jogamp/jar/gluegen-rt.jar:${ROOT_DIR}/third_party/jogamp/jar/jogl-all.jar:${ROOT_DIR}/third_party/jabel/asm-9.6.jar:${ROOT_DIR}/third_party/jabel/asm-commons-9.6.jar:${ROOT_DIR}/third_party/jabel/asm-tree-9.6.jar:${ROOT_DIR}/third_party/jabel/streamsupport-1.7.2.jar"
 
 mkdir -p "$OUT_PATH"
-"${JAVA_HOME}/bin/javadoc" --release 17 -encoding UTF-8 -docencoding UTF-8 -charset UTF-8 -Xdoclint:none -Werror -notimestamp -windowtitle "CEF Java API Docs" -bottom "<center><a href='https://github.com/chromiumembedded/java-cef' target='_top'>Chromium Embedded Framework (CEF)</a> Copyright &copy; 2013 Marshall A. Greenblatt</center>" -nodeprecated -d "$OUT_PATH" -classpath "$CLASS_PATH" -sourcepath "${ROOT_DIR}/java" -link https://docs.oracle.com/en/java/javase/17/docs/api/ -subpackages org.cef
+"${JAVA_HOME}/bin/javadoc" --release 17 -encoding UTF-8 -docencoding UTF-8 -charset UTF-8 -Xdoclint:none -Werror -notimestamp -windowtitle "CEF Java API Docs" -bottom "<center><a href='https://github.com/Keksuccino/jcef-rinku' target='_top'>JCEF for Rinku</a> Copyright &copy; 2013 Marshall A. Greenblatt</center>" -nodeprecated -d "$OUT_PATH" -classpath "$CLASS_PATH" -sourcepath "${ROOT_DIR}/java" -link https://docs.oracle.com/en/java/javase/17/docs/api/ -subpackages org.cef